+++ ado_wrapper/typing_info.py
@@ -7,7 +7,8 @@ from ado_wrapper.resources.commits import CommitChangeType
 from ado_wrapper.resources.merge_policies import WhenChangesArePushed
 from ado_wrapper.resources.permissions import PermissionGroupLiteral, PermissionActionType, PermissionType
 from ado_wrapper.resources.pull_requests import PullRequestStatus, MergeStatus, DraftState, CommentType, PrCommentStatus
-from ado_wrapper.resources.projects import ProjectVisibilityType, TemplateTypes, ProjectStatus, ProjectRepositorySettingType
+from ado_wrapper.resources.projects import ProjectVisibilityType, TemplateTypes, ProjectStatus
+from ado_wrapper.resources.project_settings import ProjectRepositorySettingType, RepoPolicyProgrammaticName, RepoPolicyDisplayTypes
 from ado_wrapper.resources.releases import ReleaseStatus
 from ado_wrapper.resources.repo_user_permission import RepoPermsActionType, RepoPermissionType
 from ado_wrapper.resources.runs import RunResult, RunState, JobStateLiteral, JobResultLiteral